Only call iommu if doing read or write.
The previous code was calling the iommu if the IORQ line
was asserted, but (apparently!) the address pins are not
valid at this time, and the iommu was dispatching to
a bogus device. (The symptom was a warning for IO access
to the unimplemented SD card $80.) Added a check to
read or write before dispatching to iommu.

Also added a debug print when writing to the ROM. The
PC and the bus pins are dumped to cout. This is to
help debug a problem where the ROM is being written
to in a part of the code that's not doing any writes.